% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_likelihood.R
\name{logLik.likelihood_function}
\alias{logLik.likelihood_function}
\title{Extract log-likelihood from a likelihood function}
\usage{
\method{logLik}{likelihood_function}(object, ...)
}
\arguments{
\item{object}{A \code{likelihood_function} object}

\item{...}{Additional arguments (unused)}
}
\value{
An object of class \code{logLik} containing the log-likelihood at the MLE,
with attributes \code{df} (number of parameters) and \code{nobs} (number of observations)
}
\description{
Extract log-likelihood from a likelihood function
}
\details{
Returns the log-likelihood evaluated at the maximum likelihood estimates.
The value includes attributes that make it compatible with information criteria
calculations (AIC, BIC, etc.).
}
