% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Graph and Compare Consonance, Surprisal, and Likelihood Functions}
\usage{
plot_compare(data1, data2, type = c("c", "s", "l1", "l2", "l3", "d"),
  measure = c("default", "ratio"), nullvalue = FALSE,
  position = c("pyramid", "inverted"), title = "Interval Functions",
  subtitle = "The function displays intervals at every level.",
  xaxis = expression(theta == ~"Range of Values"),
  yaxis1 = expression(paste(italic(p), "-value")),
  yaxis2 = "Levels for CI (\%)", color1 = darken("#D55E00", 0.2),
  color2 = darken("#009E73", 0.2), fill1 = "#99c7c7", fill2 = "#d46c5b")
}
\arguments{
\item{data1}{The first dataframe produced by one of the interval functions in which the
intervals are stored.}

\item{data2}{The second dataframe produced by one of the interval functions in which the
intervals are stored.}

\item{type}{The kind of function to plot. One of \code{"c"} for a consonance
function, the default, \code{"s"} for a surprisal function, \code{"l1"} for
relative likelihood, \code{"l2"} for log-likelihood, \code{"l3"} for likelihood,
or \code{"d"} for the deviance function. Unlike \code{\link[=ggcurve]{ggcurve()}}, \code{plot_compare()}
implements neither \code{"cdf"} nor \code{"cd"}. An unrecognized value is an error
naming the permitted ones.}

\item{measure}{Indicates whether the object has a log transformation
or is normal/default. The default setting is "default". If the measure
is set to "ratio", it will take logarithmically transformed values and
convert them back to normal values in the dataframe. This is typically a
setting used for binary outcomes and their measures such as risk ratios,
hazard ratios, and odds ratios.}

\item{nullvalue}{Indicates whether the null value for the measure
should be plotted. By default, it is set to FALSE, meaning it will not be
plotted as a vertical line. Changing this to TRUE, will plot a vertical
line at 0 when the measure is set to " default" and a vertical line at
1 when the measure is set to "ratio". For example,
plot_compare(type = "c", data = df, measure = "ratio", nullvalue = "present").
This feature is not yet available for surprisal functions.}

\item{position}{Determines the orientation of the P-value (consonance) function.
By default, it is set to "pyramid", meaning the p-value function will
stand right side up, like a pyramid. However, it can also be inverted
via the option "inverted". This will also change the sequence of the
y-axes to match the orientation.This can be set as such,
plot_compare(type = "c", data = df, position = "inverted").}

\item{title}{A custom title for the graph. By default, it is
set to "Consonance Function". In order to set a title, it must
be in quotes. For example, plot_compare(type = "c",
data = x, title = "Custom Title").}

\item{subtitle}{A custom subtitle for the graph. By default, it is set
to "The function contains consonance/confidence intervals at every level
and the P-values." In order to set a subtitle, it must be in quotes.
For example, plot_compare(type = "c", data = x, subtitle = "Custom Subtitle").}

\item{xaxis}{A custom x-axis title for the graph. By default,
it is set to "Range of Values.
In order to set a x-axis title, it must be in quotes. For example,
plot_compare(type = "c", data = x, xaxis = "Hazard Ratio").}

\item{yaxis1}{A custom y-axis title for the graph. By default,
it is set to "Consonance Level".
In order to set a y-axis title, it must be in quotes. For example,
ggcurve(type = "c", data = x, yxis1= "Confidence Level").}

\item{yaxis2}{A custom y-axis title for the graph. By default,
it is set to "Levels for CI".
In order to set a y-axis title, it must be in quotes. For example,
ggcurve(type = "c", data = x, yxis2= "Confidence Level").}

\item{color1}{Item that allows the user to choose the color of the points
and the ribbons in the graph. By default, it is set to darken("#D55E00", 0.4).
The inputs must be in quotes.}

\item{color2}{Item that allows the user to choose the color of the points
and the ribbons in the graph. By default, it is set to darken("#009E73", 0.4).
The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, color = "#333333").}

\item{fill1}{Item that allows the user to choose the color of the ribbons in the graph
for data1. By default, it is set to fill1 = "#239a98". The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, fill1 = "#333333").}

\item{fill2}{Item that allows the user to choose the color of the ribbons in the graph
for data1. By default, it is set to fill2 = "#d46c5b". The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, fill2 = "#333333").}
}
\value{
A plot that compares two functions.
}
\description{
Compares the p-value/s-value, and likelihood functions using ggplot2 graphics.
}
\examples{
# How far does adjusting for horsepower move the estimated effect of
# weight on fuel economy? Overlaying the two curves shows the shift and
# the change in precision at every level at once.
unadjusted <- curve_gen(lm(mpg ~ wt, data = mtcars), "wt")
adjusted <- curve_gen(lm(mpg ~ wt + hp, data = mtcars), "wt")

plot_compare(unadjusted[[1]], adjusted[[1]], type = "c", nullvalue = TRUE)

}
\seealso{
\code{\link[=ggcurve]{ggcurve()}}

\code{\link[=curve_compare]{curve_compare()}}
}
