% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{RobustMax}
\alias{RobustMax}
\title{Robust Max, an alternative to max() that doesn't throw a warning}
\usage{
RobustMax(x)
}
\arguments{
\item{x}{A vector to find the maximum value of}
}
\value{
The max value from a vector
}
\description{
Robust Max, an alternative to max() that doesn't throw a warning
}
