% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{RobustMin}
\alias{RobustMin}
\title{Robust Min, an alternative to max() that doesn't throw a warning}
\usage{
RobustMin(x)
}
\arguments{
\item{x}{A vector find the minimum value of}
}
\value{
The minimum value from the vector
}
\description{
Robust Min, an alternative to max() that doesn't throw a warning
}
