% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_gen.R
\name{curve_gen}
\alias{curve_gen}
\title{Computes thousands of consonance (confidence) intervals for the chosen parameter in the}
\usage{
curve_gen(
  model,
  var,
  method = "wald",
  replicates = 1000,
  steps = 10000,
  table = TRUE,
  ...
)
}
\arguments{
\item{model}{The statistical model of interest(ANOVA, regression, logistic regression) is to be indicated here.}

\item{var}{The variable of interest from the model (coefficients, intercept) for which the intervals are to be produced.}

\item{method}{Chooses the method to be used to calculate the consonance intervals. There are currently four}

\item{replicates}{Indicates how many bootstrap replicates are to be performed if bootstrapping is enabled as a method.}

\item{steps}{Indicates how many consonance intervals are to be calculated at various levels.
For example, setting this to 100 will produce 100 compatibility intervals from 0 to 100.
Setting this to 10000 will produce more consonance levels. By default, it is set to 1000.
Increasing the number substantially is not recommended as it will take longer to produce all
the intervals and store them into a dataframe.}

\item{table}{Indicates whether or not a table output with some relevant statistics should}
}
\value{

}
\description{
Computes thousands of consonance (confidence) intervals for the chosen parameter in the
}
\examples{
# Simulate random data
GroupA <- rnorm(50)
GroupB <- rnorm(50)
RandomData <- data.frame(GroupA, GroupB)
rob <- glm(GroupA ~ GroupB, data = RandomData)
bob <- curve_gen(rob, "GroupB", method = "lm")
tibble::tibble(bob)
}
