% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_lik.R
\name{curve_lik}
\alias{curve_lik}
\title{Compute the Profile Likelihood Functions}
\usage{
curve_lik(likobject, data, table = TRUE)
}
\arguments{
\item{likobject}{An object from the ProfileLikelihood package}

\item{data}{The dataframe that was used to create the likelihood object in the ProfileLikelihood package.}

\item{table}{Indicates whether or not a table output with some relevant statistics should}
}
\description{
Compute the Profile Likelihood Functions
}
\examples{

data(dataglm)
xx <- profilelike.glm(y ~ x1 + x2, data = dataglm, profile.theta = "group", family = binomial(link = "logit"), length = 500, round = 2)
(curve_lik(xx, dataglm))
}
