\name{defunct}
\alias{defunct}
\alias{plotpint}
\alias{plotsint}
\alias{meanintervals}
\alias{metaintervals}
\alias{genintervals}
\alias{corrintervals}
\alias{survintervals}
\alias{likintervals}
\alias{rev_eng}
\alias{ggconcurve}
\alias{plot_concurve}
\title{Deprecated functions in \pkg{concurve}.}
\usage{
plotpint(...)

plotsint(...)

meanintervals(...)

metaintervals(...)

genintervals(...)

corrintervals(...)

survintervals(...)

rev_eng(...)

ggconcurve(...)

plot_concurve(...)

}

\description{
The functions listed below are deprecated. Please use the listed alternatives.}

\section{\code{plotpint}}{

For \code{plotpint}, use \code{\link{ggcurve}}.
}

\section{\code{plotsint}}{

For \code{plotsint}, use \code{\link{ggcurve}}.
}

\section{\code{plot_concurve}}{

For \code{plot_concurve}, use \code{\link{ggcurve}}.
}

\section{\code{ggconcurve}}{

For \code{ggconcurve}, use \code{\link{ggcurve}}.
}

\section{\code{meanintervals}}{

For \code{meanintervals}, use \code{\link{curve_mean}}.
}

\section{\code{metaintervals}}{

For \code{metaintervals}, use \code{\link{curve_meta}}.
}

\section{\code{genintervals}}{

For \code{genintervals}, use \code{\link{curve_gen}}.
}

\section{\code{corrintervals}}{

For \code{corrintervals}, use \code{\link{curve_corr}}.
}

\section{\code{survintervals}}{

For \code{survintervals}, use \code{\link{curve_surv}}.
}

\section{\code{rev_eng}}{

For \code{rev_eng}, use \code{\link{curve_rev}}.
}

\keyword{internal}

