% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare.R
\name{plot_compare}
\alias{plot_compare}
\title{Graph and Compare Consonance, Surprisal, and Likelihood Functions}
\usage{
plot_compare(data1, data2, type = "c", measure = "default",
  nullvalue = FALSE, position = "pyramid", title = "Interval Functions",
  subtitle = "The function displays intervals at every level.",
  xaxis = expression(Theta ~ "Range of Values"),
  yaxis = expression(paste(italic(p), "-value")), color = "#000000",
  fill1 = "#239a98", fill2 = "#EE6A50")
}
\arguments{
\item{data1}{The first dataframe produced by one of the interval functions in which the
intervals are stored.}

\item{data2}{The second dataframe produced by one of the interval functions in which the
intervals are stored.}

\item{type}{Choose whether to plot a "consonance" function, a
"surprisal" function or "likelihood". The default option is set to "c".
The type must be set in quotes, for example plot_compare(type = "s") or
plot_compare(type = "c"). Other options include "pd" for the consonance
distribution function, and "cd" for the consonance density function,
"l1" for relative likelihood, "l2" for log-likelihood, "l3" for likelihood
and "d" for deviance function.}

\item{measure}{Indicates whether the object has a log transformation
or is normal/default. The default setting is "default". If the measure
is set to "ratio", it will take logarithmically transformed values and
convert them back to normal values in the dataframe. This is typically a
setting used for binary outcomes and their measures such as risk ratios,
hazard ratios, and odds ratios.}

\item{nullvalue}{Indicates whether the null value for the measure
should be plotted. By default, it is set to FALSE, meaning it will not be
plotted as a vertical line. Changing this to TRUE, will plot a vertical
line at 0 when the measure is set to " default" and a vertical line at
1 when the measure is set to "ratio". For example,
plot_compare(type = "c", data = df, measure = "ratio", nullvalue = "present").
This feature is not yet available for surprisal functions.}

\item{position}{Determines the orientation of the P-value (consonance) function.
By default, it is set to "pyramid", meaning the p-value function will
stand right side up, like a pyramid. However, it can also be inverted
via the option "inverted". This will also change the sequence of the
y-axes to match the orientation.This can be set as such,
plot_compare(type = "c", data = df, position = "inverted").}

\item{title}{A custom title for the graph. By default, it is
set to "Consonance Function". In order to set a title, it must
be in quotes. For example, plot_compare(type = "c",
data = x, title = "Custom Title").}

\item{subtitle}{A custom subtitle for the graph. By default, it is set
to "The function contains consonance/confidence intervals at every level
and the P-values." In order to set a subtitle, it must be in quotes.
For example, plot_compare(type = "c", data = x, subtitle = "Custom Subtitle").}

\item{xaxis}{A custom x-axis title for the graph. By default,
it is set to "Range of Values.
In order to set a x-axis title, it must be in quotes. For example,
plot_compare(type = "c", data = x, xaxis = "Hazard Ratio").}

\item{yaxis}{A custom y-axis title for the graph. By default,
it is set to "Consonance Level".
In order to set a y-axis title, it must be in quotes. For example,
plot_compare(type = "c", data = x, yxis = "Confidence Level").}

\item{color}{Item that allows the user to choose the color of the points
and the ribbons in the graph. By default, it is set to color = "#555555".
The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, color = "#333333").}

\item{fill1}{Item that allows the user to choose the color of the ribbons in the graph
for data1. By default, it is set to fill1 = "#239a98". The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, fill1 = "#333333").}

\item{fill2}{Item that allows the user to choose the color of the ribbons in the graph
for data1. By default, it is set to fill2 = "#d46c5b". The inputs must be in quotes.
For example, plot_compare(type = "c", data = x, fill2 = "#333333").}
}
\value{
A plot that compares two functions.
}
\description{
Compares the p-value/s-value, and likelihood functions using ggplot2 graphics.
}
\examples{
\donttest{
library(concurve)

GroupA <- rnorm(50)
GroupB <- rnorm(50)
RandomData <- data.frame(GroupA, GroupB)
intervalsdf <- curve_mean(GroupA, GroupB, data = RandomData)
GroupA2 <- rnorm(50)
GroupB2 <- rnorm(50)
RandomData2 <- data.frame(GroupA2, GroupB2)
model <- lm(GroupA2 ~ GroupB2, data = RandomData2)

randomframe <- curve_gen(model, "GroupB2")

plot_compare(intervalsdf[[1]], randomframe[[1]], type = "c")
}

}
\references{
Birnbaum, A. (1961), “A unified theory of estimation, I,” The Annals of Mathematical Statistics, 32, 112–135. https://doi.org/10.1214/aoms/1177705145.

Chow, Z. R., and Greenland, S. (2019), “Semantic and Cognitive Tools to Aid Statistical Inference: Replace Confidence and Significance by Compatibility and Surprise,” arXiv:1909.08579.

Fraser, D. A. S. (2017), “p-Values: The Insight to Modern Statistical Inference,” Annual Review of Statistics and Its Application, 4, 1–14. https://doi.org/10.1146/annurev-statistics-060116-054139.

Fraser, D. A. S. (2019), “The P-value function and statistical inference,” The American Statistician, 73, 135–147. https://doi.org/10.1080/00031305.2018.1556735.

Hjort, N. L., and Schweder, T. (2018), “Confidence distributions and related themes,” Journal of Statistical Planning and Inference, 195, 1–13. https://doi.org/10.1016/j.jspi.2017.09.017.

Infanger, D., and Schmidt‐Trucksäss, A. (2019), “P value functions: An underused method to present research results and to promote quantitative reasoning,” Statistics in Medicine. https://doi.org/10.1002/sim.8293.

Poole, C. (1987a), “Beyond the confidence interval,” American Journal of Public Health, 77, 195–199. https://doi.org/10.2105/AJPH.77.2.195.

Poole, C. (1987b), “Confidence intervals exclude nothing.,” American Journal of Public Health, 77, 492–493. https://doi.org/10.2105/ajph.77.4.492.

Rothman, K. J., Greenland, S., and Lash, T. L. (2008a), Modern Epidemiology, Lippincott Williams & Wilkins.

Rothman, K. J., Greenland, S., and Lash, T. L. (2008b), “Precision and statistics in epidemiologic studies,” in Modern Epidemiology, eds. K. J. Rothman, S. Greenland, and T. L. Lash, Lippincott Williams & Wilkins, pp. 148–167.

Schweder, T., and Hjort, N. L. (2002), “Confidence and Likelihood*,” Scandinavian Journal of Statistics, 29, 309–332. https://doi.org/10.1111/1467-9469.00285.

Schweder, T., and Hjort, N. L. (2016), Confidence, Likelihood, Probability: Statistical Inference with Confidence Distributions, Cambridge University Press.

Sullivan, K. M., and Foster, D. A. (1990), “Use of the confidence interval function,” Epidemiology, 1, 39–42. https://doi.org/10.1097/00001648-199001000-00009.

Xie, M., and Singh, K. (2013), “Confidence Distribution, the Frequentist Distribution Estimator of a Parameter: A Review,” International Statistical Review, 81, 3–39. https://doi.org/10.1111/insr.12000.
}
\seealso{
\code{\link{ggcurve}}

\code{\link{curve_compare}}
}
